Allow login with email (#239)
While the call to obtain a JWT (`api/v3/user/login`) works for username or email, the app was then following that with a call to ` api/v3/user`, which only works with username (https://lemmy.readme.io/reference/get_user). This would cause the login attempt to fail with the error reported on #235.

These changes skip the call to `api/v3/user`, and instead retrieves the user information from the call to `api/v3/site`, that was already being performed.

Fixes #235.

Question: should the label of the username field be updated to "Username / Email" or "Username or Email"? (screenshot)

Reviewed-on: https://codeberg.org/Bazsalanszky/Eternity/pulls/239
Reviewed-by: Bazsalanszky <bazsalanszky@noreply.codeberg.org>
Co-authored-by: tinsukE <tinsuke@noreply.codeberg.org>
Co-committed-by: tinsukE <tinsuke@noreply.codeberg.org>